Fixed Sound Emmiter Light Issue
I switched to storing a soundEmmiter's light as an index value in the scene's list of lights, instead of pointers like they were before. This comes at the price of not being able to easily delete lights at run-time, but we could easily make a function that shifts over sound emmiter's index values when needed.